meeting-summarizer/requirements.txt
python-dotenv>=0.19.0
openai>=0.27.0